requirements.txt
langchain
openai
langchain-community
huggingface_hub
python-dotenv
streamlit
openai